requirements.txt
asgiref==3.2.10
astroid==2.4.2
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
checkout==0.2.0
dateparser==1.0.0
Django==2.2.6
django-cors-headers==3.6.0
django-debug-toolbar==3.2
django-environ==0.4.5
django-extensions==3.1.0
django-filter==2.4.0
djangorestframework==3.11.0
djangorestframework-filters==0.3.0
djangorestframework-simplejwt==4.3.0
export==0.1.2
Faker==6.0.0
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.7.0
iniconfig==1.1.1
isort==5.6.4
lark-parser==0.11.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.2.0
ordered-set==4.0.2
packaging==20.1
pbr==5.5.1
Pillow==7.0.0
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.8.1
PyJWT==1.7.1
pylint==2.6.0
pyparsing==2.4.6
pytest==5.3.5
pytest-django==3.8.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.4.1
recipe==0.23.0
recipes==0.1
regex==2020.11.13
requests==2.22.0
six==1.14.0
sorl-thumbnail==12.6.3
SQLAlchemy==1.3.23
sqlparse==0.3.0
stevedore==3.3.0
sureberus==0.14.0
tablib==3.0.0
text-unidecode==1.3
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.6
username==2.0.2
wcwidth==0.1.8
wrapt==1.12.1
zipp==2.2.0
